Virtualization: include scrollbar width in calculation
