+++ public/app/features/logs/components/panel/LogLine.tsx
@@ -1,11 +1,10 @@
 import { css } from '@emotion/css';
 import { CSSProperties } from 'react';
 
-import { GrafanaTheme2, LogRowModel } from '@grafana/data';
+import { GrafanaTheme2 } from '@grafana/data';
 import { useTheme2 } from '@grafana/ui';
 
 import { ProcessedLogModel } from './processing';
-import { measureText } from './virtualization';
 
 interface Props {
   log: ProcessedLogModel;
@@ -42,21 +41,3 @@ const getStyles = (theme: GrafanaTheme2) => ({
     whiteSpace: 'pre',
   }),
 });
-
-export function getLogLineSize(
-  logs: ProcessedLogModel[],
-  container: HTMLDivElement | null,
-  theme: GrafanaTheme2,
-  wrapLogMessage: boolean,
-  index: number
-) {
-  if (!container) {
-    return 0;
-  }
-  const lineHeight = theme.typography.fontSize * theme.typography.body.lineHeight;
-  if (!wrapLogMessage) {
-    return lineHeight;
-  }
-  const { height } = measureText(logs[index].body, container.clientWidth, lineHeight);
-  return height;
-}
